% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{download.ClimateData}
\alias{download.ClimateData}
\title{Base utility for Climate Bulk Data fetching}
\usage{
download.ClimateData(station, begin, end, timeframe = c("hourly",
  "daily", "monthly"), method = c("curl", "wget"))
}
\arguments{
\item{method}{}
}
\value{
data.frame with numeral vectors where possible
}
\description{
Base utility for Climate Bulk Data fetching
}
\examples{
\donttest{
# Station 51423 is KAMLOOPS A
begin <- 2014
end <- 2017
df <- download.ClimateData(51423, begin, end, "hourly", "curl")
}
}
