% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{download.ClimateDataAll}
\alias{download.ClimateDataAll}
\title{Download all available data for the specified \code{timeframe}}
\usage{
download.ClimateDataAll(timeframe)
}
\arguments{
\item{timeframe}{of station/datastream}
}
\description{
See \code{\link{download.ClimateData}} for available timeframes
}
