% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{download.ClimateDataStations}
\alias{download.ClimateDataStations}
\title{Download all data from \code{stations} for the specified \code{timeframe}}
\usage{
download.ClimateDataStations(stations, timeframe)
}
\arguments{
\item{stations}{to download}

\item{timeframe}{of stations/datastreams}
}
\description{
See \code{\link{download.ClimateData}} for available timeframes
See \code{stations.csv} (\code{\link{read.ClimateStationList}}) for available Station IDs
}
