% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.Climate}
\alias{plot.Climate}
\title{Plot the temperature on a time-series graph}
\usage{
\method{plot}{Climate}(station, years = NA, begin = NA, end = NA,
  mode = c("MD", "Date"))
}
\arguments{
\item{station}{to plot}

\item{years}{to plot (can be filled in by begin or end)}

\item{begin}{year (opt)}

\item{end}{year (opt)}

\item{mode}{MD or Date}
}
\description{
Plot the temperature on a time-series graph
}
