% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{read.ClimateCSV}
\alias{read.ClimateCSV}
\title{Read the downloaded data for \code{station}, \code{year}}
\usage{
read.ClimateCSV(station, year)
}
\arguments{
\item{station}{to read data}

\item{year}{of compiled csv}
}
\description{
Must run \code{\link{compile.ClimateData}} first
}
