% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{read.ClimateHeader}
\alias{read.ClimateHeader}
\title{Read the header information for \code{station}}
\usage{
read.ClimateHeader(station)
}
\arguments{
\item{station}{to load header}
}
\description{
Read the header information for \code{station}
}
