% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Util.R
\name{download.ClimateData}
\alias{download.ClimateData}
\title{Base utility for Bulk Climate Data fetching}
\usage{
download.ClimateData(station, begin, end, timeframe = c("hourly",
  "daily", "monthly"), method = c("curl", "wget"))
}
\arguments{
\item{station}{to download}

\item{begin}{year}

\item{end}{year}

\item{timeframe}{of station/datastream}

\item{method}{curl or wget}
}
\description{
Download climate data from \code{begin}:\code{end}
\code{station} is provided by Station.ID in stations.csv
\code{timeframe} is either \code{"hourly"}, \code{"daily"}, or \code{"monthly"}
}
\details{
wget is the \emph{supported} utility but curl works fine
and is more accessible on all systems
}
\examples{
\donttest{
# Station 51423 is KAMLOOPS A
begin <- 2014
end <- 2017
df <- download.ClimateData(51423, begin, end, "hourly", "curl")
}
}
\references{
\url{ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Readme.txt}
\url{ftp://ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Station\%20Inventory\%20EN.csv}
}
